removed redunant channels class / file; refactored channels into it's own python module; WIP for providing formatting based on channel type; WIP for normalizing channel types
