+++ autosportlabs/racecapture/databus/databus.py
@@ -1,8 +1,8 @@
 from kivy.clock import Clock
 from time import sleep
 from threading import Thread, Event
-from autosportlabs.racecapture.data.sampledata import Sample, ChannelMeta, SampleMetaException,\
-    ChannelMetaCollection
+from autosportlabs.racecapture.data.channels import ChannelMeta
+from autosportlabs.racecapture.data.sampledata import Sample, SampleMetaException, ChannelMetaCollection
 from autosportlabs.racecapture.databus.filter.bestlapfilter import BestLapFilter
 from autosportlabs.racecapture.databus.filter.laptimedeltafilter import LaptimeDeltaFilter
 from autosportlabs.racecapture.databus.filter.currentlaptimefilter import CurrentLapTimeFilter
